requirements.txt
fastapi
uvicorn
python-dotenv
weaviate-client
google-generativeai
pymupdf
camelot-py[cv]
pandas
tabulate